Adds ViewBinding Adds the Navigation dependencies Implements navigation with bottom navigation Adds the app's defined colors Adds the icon drawables for the selected and unselected states
Fixed #1
