+++ app/src/main/java/com/lucassimao/movielanddibbarbosa/feature/home/MoviesFragment.kt
@@ -1,10 +1,11 @@
-package com.lucassimao.movielanddibbarbosa
+package com.lucassimao.movielanddibbarbosa.feature.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.lucassimao.movielanddibbarbosa.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
